Refactored code into modular components
