+++ .gitignore
@@ -1,0 +1,3 @@
+.cache
+build
+build-asan
