+++ build-asan/CMakeFiles/bob16.dir/build.make
@@ -86,18 +86,50 @@ CMakeFiles/bob16.dir/src/main.c.s: cmake_force
 	@$(CMAKE_COMMAND) -E cmake_echo_color "--switch=$(COLOR)" --green "Compiling C source to assembly CMakeFiles/bob16.dir/src/main.c.s"
 	/usr/bin/cc $(C_DEFINES) $(C_INCLUDES) $(C_FLAGS) -S /home/Auth/Projects/bob16/src/main.c -o CMakeFiles/bob16.dir/src/main.c.s
 
+CMakeFiles/bob16.dir/src/cpu.c.o: CMakeFiles/bob16.dir/flags.make
+CMakeFiles/bob16.dir/src/cpu.c.o: /home/Auth/Projects/bob16/src/cpu.c
+CMakeFiles/bob16.dir/src/cpu.c.o: CMakeFiles/bob16.dir/compiler_depend.ts
+	@$(CMAKE_COMMAND) -E cmake_echo_color "--switch=$(COLOR)" --green --progress-dir=/home/Auth/Projects/bob16/build-asan/CMakeFiles --progress-num=$(CMAKE_PROGRESS_2) "Building C object CMakeFiles/bob16.dir/src/cpu.c.o"
+	/usr/bin/cc $(C_DEFINES) $(C_INCLUDES) $(C_FLAGS) -MD -MT CMakeFiles/bob16.dir/src/cpu.c.o -MF CMakeFiles/bob16.dir/src/cpu.c.o.d -o CMakeFiles/bob16.dir/src/cpu.c.o -c /home/Auth/Projects/bob16/src/cpu.c
+
+CMakeFiles/bob16.dir/src/cpu.c.i: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color "--switch=$(COLOR)" --green "Preprocessing C source to CMakeFiles/bob16.dir/src/cpu.c.i"
+	/usr/bin/cc $(C_DEFINES) $(C_INCLUDES) $(C_FLAGS) -E /home/Auth/Projects/bob16/src/cpu.c > CMakeFiles/bob16.dir/src/cpu.c.i
+
+CMakeFiles/bob16.dir/src/cpu.c.s: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color "--switch=$(COLOR)" --green "Compiling C source to assembly CMakeFiles/bob16.dir/src/cpu.c.s"
+	/usr/bin/cc $(C_DEFINES) $(C_INCLUDES) $(C_FLAGS) -S /home/Auth/Projects/bob16/src/cpu.c -o CMakeFiles/bob16.dir/src/cpu.c.s
+
+CMakeFiles/bob16.dir/src/assembler.c.o: CMakeFiles/bob16.dir/flags.make
+CMakeFiles/bob16.dir/src/assembler.c.o: /home/Auth/Projects/bob16/src/assembler.c
+CMakeFiles/bob16.dir/src/assembler.c.o: CMakeFiles/bob16.dir/compiler_depend.ts
+	@$(CMAKE_COMMAND) -E cmake_echo_color "--switch=$(COLOR)" --green --progress-dir=/home/Auth/Projects/bob16/build-asan/CMakeFiles --progress-num=$(CMAKE_PROGRESS_3) "Building C object CMakeFiles/bob16.dir/src/assembler.c.o"
+	/usr/bin/cc $(C_DEFINES) $(C_INCLUDES) $(C_FLAGS) -MD -MT CMakeFiles/bob16.dir/src/assembler.c.o -MF CMakeFiles/bob16.dir/src/assembler.c.o.d -o CMakeFiles/bob16.dir/src/assembler.c.o -c /home/Auth/Projects/bob16/src/assembler.c
+
+CMakeFiles/bob16.dir/src/assembler.c.i: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color "--switch=$(COLOR)" --green "Preprocessing C source to CMakeFiles/bob16.dir/src/assembler.c.i"
+	/usr/bin/cc $(C_DEFINES) $(C_INCLUDES) $(C_FLAGS) -E /home/Auth/Projects/bob16/src/assembler.c > CMakeFiles/bob16.dir/src/assembler.c.i
+
+CMakeFiles/bob16.dir/src/assembler.c.s: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color "--switch=$(COLOR)" --green "Compiling C source to assembly CMakeFiles/bob16.dir/src/assembler.c.s"
+	/usr/bin/cc $(C_DEFINES) $(C_INCLUDES) $(C_FLAGS) -S /home/Auth/Projects/bob16/src/assembler.c -o CMakeFiles/bob16.dir/src/assembler.c.s
+
 # Object files for target bob16
 bob16_OBJECTS = \
-"CMakeFiles/bob16.dir/src/main.c.o"
+"CMakeFiles/bob16.dir/src/main.c.o" \
+"CMakeFiles/bob16.dir/src/cpu.c.o" \
+"CMakeFiles/bob16.dir/src/assembler.c.o"
 
 # External object files for target bob16
 bob16_EXTERNAL_OBJECTS =
 
 bob16: CMakeFiles/bob16.dir/src/main.c.o
+bob16: CMakeFiles/bob16.dir/src/cpu.c.o
+bob16: CMakeFiles/bob16.dir/src/assembler.c.o
 bob16: CMakeFiles/bob16.dir/build.make
 bob16: CMakeFiles/bob16.dir/compiler_depend.ts
 bob16: CMakeFiles/bob16.dir/link.txt
-	@$(CMAKE_COMMAND) -E cmake_echo_color "--switch=$(COLOR)" --green --bold --progress-dir=/home/Auth/Projects/bob16/build-asan/CMakeFiles --progress-num=$(CMAKE_PROGRESS_2) "Linking C executable bob16"
+	@$(CMAKE_COMMAND) -E cmake_echo_color "--switch=$(COLOR)" --green --bold --progress-dir=/home/Auth/Projects/bob16/build-asan/CMakeFiles --progress-num=$(CMAKE_PROGRESS_4) "Linking C executable bob16"
 	$(CMAKE_COMMAND) -E cmake_link_script CMakeFiles/bob16.dir/link.txt --verbose=$(VERBOSE)
 
 # Rule to build all files generated by this target.
