+++ build-asan/CMakeFiles/bob16.dir/compiler_depend.internal
@@ -1,8 +1,9 @@
 # CMAKE generated file: DO NOT EDIT!
 # Generated by "Unix Makefiles" Generator, CMake Version 4.4
 
-CMakeFiles/bob16.dir/src/main.c.o
- /home/Auth/Projects/bob16/src/main.c
+CMakeFiles/bob16.dir/src/cpu.c.o
+ /home/Auth/Projects/bob16/src/cpu.c
+ /home/Auth/Projects/bob16/include/cpu.h
  /usr/include/alloca.h
  /usr/include/bits/atomic_wide_counter.h
  /usr/include/bits/byteswap.h
@@ -67,4 +68,48 @@ CMakeFiles/bob16.dir/src/main.c.o
  /usr/lib/gcc/x86_64-pc-linux-gnu/16/include/stdbool.h
  /usr/lib/gcc/x86_64-pc-linux-gnu/16/include/stddef.h
  /usr/lib/gcc/x86_64-pc-linux-gnu/16/include/stdint.h
+
+CMakeFiles/bob16.dir/src/main.c.o
+ /home/Auth/Projects/bob16/src/main.c
+ /home/Auth/Projects/bob16/include/assembler.h
+ /home/Auth/Projects/bob16/include/cpu.h
+ /usr/include/bits/libc-header-start.h
+ /usr/include/bits/long-double.h
+ /usr/include/bits/stdint-intn.h
+ /usr/include/bits/stdint-least.h
+ /usr/include/bits/stdint-uintn.h
+ /usr/include/bits/time64.h
+ /usr/include/bits/timesize.h
+ /usr/include/bits/types.h
+ /usr/include/bits/typesizes.h
+ /usr/include/bits/wchar.h
+ /usr/include/bits/wordsize.h
+ /usr/include/features-time64.h
+ /usr/include/features.h
+ /usr/include/gnu/stubs-64.h
+ /usr/include/gnu/stubs.h
+ /usr/include/stdc-predef.h
+ /usr/include/stdint.h
+ /usr/include/sys/cdefs.h
+ /usr/lib/gcc/x86_64-pc-linux-gnu/16/include/stdbool.h
+ /usr/lib/gcc/x86_64-pc-linux-gnu/16/include/stdint.h
+
+bob16
+ /usr/lib/Scrt1.o
+ /usr/lib/crti.o
+ /usr/lib/crtn.o
+ /usr/lib/libc.so
+ /usr/lib/libgcc_s.so.1
+ /usr/lib/gcc/x86_64-pc-linux-gnu/16/crtbeginS.o
+ /usr/lib/gcc/x86_64-pc-linux-gnu/16/crtendS.o
+ /usr/lib/gcc/x86_64-pc-linux-gnu/16/libatomic.so
+ /usr/lib/gcc/x86_64-pc-linux-gnu/16/libatomic_asneeded.so
+ /usr/lib/gcc/x86_64-pc-linux-gnu/16/libgcc.a
+ /usr/lib/gcc/x86_64-pc-linux-gnu/16/libgcc_s.so
+ /usr/lib/gcc/x86_64-pc-linux-gnu/16/libgcc_s_asneeded.so
+ /usr/lib/ld-linux-x86-64.so.2
+ /usr/lib/libc.so.6
+ /usr/lib/libc_nonshared.a
+ /home/Auth/Projects/bob16/build-asan/CMakeFiles/bob16.dir/src/cpu.c.o
+ /home/Auth/Projects/bob16/build-asan/CMakeFiles/bob16.dir/src/main.c.o
