+++ build-asan/CMakeFiles/bob16.dir/compiler_depend.make
@@ -1,7 +1,8 @@
 # CMAKE generated file: DO NOT EDIT!
 # Generated by "Unix Makefiles" Generator, CMake Version 4.4
 
-CMakeFiles/bob16.dir/src/main.c.o: /home/Auth/Projects/bob16/src/main.c \
+CMakeFiles/bob16.dir/src/cpu.c.o: /home/Auth/Projects/bob16/src/cpu.c \
+  /home/Auth/Projects/bob16/include/cpu.h \
   /usr/include/alloca.h \
   /usr/include/bits/atomic_wide_counter.h \
   /usr/include/bits/byteswap.h \
@@ -67,6 +68,82 @@ CMakeFiles/bob16.dir/src/main.c.o: /home/Auth/Projects/bob16/src/main.c \
   /usr/lib/gcc/x86_64-pc-linux-gnu/16/include/stddef.h \
   /usr/lib/gcc/x86_64-pc-linux-gnu/16/include/stdint.h
 
+CMakeFiles/bob16.dir/src/main.c.o: /home/Auth/Projects/bob16/src/main.c \
+  /home/Auth/Projects/bob16/include/assembler.h \
+  /home/Auth/Projects/bob16/include/cpu.h \
+  /usr/include/bits/libc-header-start.h \
+  /usr/include/bits/long-double.h \
+  /usr/include/bits/stdint-intn.h \
+  /usr/include/bits/stdint-least.h \
+  /usr/include/bits/stdint-uintn.h \
+  /usr/include/bits/time64.h \
+  /usr/include/bits/timesize.h \
+  /usr/include/bits/types.h \
+  /usr/include/bits/typesizes.h \
+  /usr/include/bits/wchar.h \
+  /usr/include/bits/wordsize.h \
+  /usr/include/features-time64.h \
+  /usr/include/features.h \
+  /usr/include/gnu/stubs-64.h \
+  /usr/include/gnu/stubs.h \
+  /usr/include/stdc-predef.h \
+  /usr/include/stdint.h \
+  /usr/include/sys/cdefs.h \
+  /usr/lib/gcc/x86_64-pc-linux-gnu/16/include/stdbool.h \
+  /usr/lib/gcc/x86_64-pc-linux-gnu/16/include/stdint.h
+
+bob16: /usr/lib/Scrt1.o \
+  /usr/lib/crti.o \
+  /usr/lib/crtn.o \
+  /usr/lib/libc.so \
+  /usr/lib/libgcc_s.so.1 \
+  /usr/lib/gcc/x86_64-pc-linux-gnu/16/crtbeginS.o \
+  /usr/lib/gcc/x86_64-pc-linux-gnu/16/crtendS.o \
+  /usr/lib/gcc/x86_64-pc-linux-gnu/16/libatomic.so \
+  /usr/lib/gcc/x86_64-pc-linux-gnu/16/libatomic_asneeded.so \
+  /usr/lib/gcc/x86_64-pc-linux-gnu/16/libgcc.a \
+  /usr/lib/gcc/x86_64-pc-linux-gnu/16/libgcc_s.so \
+  /usr/lib/gcc/x86_64-pc-linux-gnu/16/libgcc_s_asneeded.so \
+  /usr/lib/ld-linux-x86-64.so.2 \
+  /usr/lib/libc.so.6 \
+  /usr/lib/libc_nonshared.a \
+  CMakeFiles/bob16.dir/src/cpu.c.o \
+  CMakeFiles/bob16.dir/src/main.c.o
+
+
+CMakeFiles/bob16.dir/src/main.c.o:
+
+CMakeFiles/bob16.dir/src/cpu.c.o:
+
+/usr/lib/libc.so.6:
+
+/usr/lib/ld-linux-x86-64.so.2:
+
+/usr/lib/gcc/x86_64-pc-linux-gnu/16/libgcc_s.so:
+
+/usr/lib/gcc/x86_64-pc-linux-gnu/16/libgcc.a:
+
+/usr/lib/gcc/x86_64-pc-linux-gnu/16/libatomic.so:
+
+/usr/lib/gcc/x86_64-pc-linux-gnu/16/libgcc_s_asneeded.so:
+
+/usr/lib/gcc/x86_64-pc-linux-gnu/16/crtendS.o:
+
+/usr/lib/gcc/x86_64-pc-linux-gnu/16/crtbeginS.o:
+
+/usr/lib/libgcc_s.so.1:
+
+/usr/lib/libc.so:
+
+/usr/lib/gcc/x86_64-pc-linux-gnu/16/libatomic_asneeded.so:
+
+/usr/lib/crtn.o:
+
+/usr/lib/crti.o:
+
+/usr/lib/Scrt1.o:
+
+/home/Auth/Projects/bob16/src/main.c:
 
 /usr/lib/gcc/x86_64-pc-linux-gnu/16/include/stdint.h:
 
@@ -78,13 +155,13 @@ CMakeFiles/bob16.dir/src/main.c.o: /home/Auth/Projects/bob16/src/main.c \
 
 /usr/include/sys/select.h:
 
+/usr/include/sys/cdefs.h:
+
 /usr/include/bits/types/FILE.h:
 
-/usr/include/bits/types/__locale_t.h:
+/usr/lib/libc_nonshared.a:
 
 /usr/include/bits/types.h:
-
-/usr/include/sys/cdefs.h:
 
 /usr/include/bits/byteswap.h:
 
@@ -104,7 +181,13 @@ CMakeFiles/bob16.dir/src/main.c.o: /home/Auth/Projects/bob16/src/main.c \
 
 /usr/include/bits/stdint-intn.h:
 
-/home/Auth/Projects/bob16/src/main.c:
+/usr/include/bits/struct_mutex.h:
+
+/usr/include/gnu/stubs-64.h:
+
+/home/Auth/Projects/bob16/include/cpu.h:
+
+/usr/include/features.h:
 
 /usr/include/bits/libc-header-start.h:
 
@@ -118,15 +201,15 @@ CMakeFiles/bob16.dir/src/main.c.o: /home/Auth/Projects/bob16/src/main.c \
 
 /usr/include/bits/waitflags.h:
 
-/usr/include/bits/struct_mutex.h:
-
-/usr/include/gnu/stubs-64.h:
-
 /usr/include/bits/endianness.h:
 
 /usr/include/bits/types/__FILE.h:
 
 /usr/include/bits/stdlib-float.h:
+
+/home/Auth/Projects/bob16/include/assembler.h:
+
+/home/Auth/Projects/bob16/src/cpu.c:
 
 /usr/include/bits/types/sigset_t.h:
 
@@ -145,6 +228,8 @@ CMakeFiles/bob16.dir/src/main.c.o: /home/Auth/Projects/bob16/src/main.c \
 /usr/include/bits/floatn.h:
 
 /usr/include/bits/pthreadtypes.h:
+
+/usr/include/bits/types/__locale_t.h:
 
 /usr/lib/gcc/x86_64-pc-linux-gnu/16/include/stddef.h:
 
@@ -187,8 +272,6 @@ CMakeFiles/bob16.dir/src/main.c.o: /home/Auth/Projects/bob16/src/main.c \
 /usr/include/endian.h:
 
 /usr/include/bits/waitstatus.h:
-
-/usr/include/features.h:
 
 /usr/include/gnu/stubs.h:
