+++ build-asan/CMakeFiles/bob16.dir/flags.make
@@ -4,7 +4,7 @@
 # compile C with /usr/bin/cc
 C_DEFINES = 
 
-C_INCLUDES = 
+C_INCLUDES = -I/home/Auth/Projects/bob16/include
 
 C_FLAGS = -std=gnu11 -Wall -Wextra -Wpedantic -Werror
