+++ build-asan/CMakeFiles/bob16.dir/link.d
@@ -3,6 +3,8 @@ bob16: \
   /usr/lib/gcc/x86_64-pc-linux-gnu/16/../../../../lib/crti.o \
   /usr/lib/gcc/x86_64-pc-linux-gnu/16/crtbeginS.o \
   CMakeFiles/bob16.dir/src/main.c.o \
+  CMakeFiles/bob16.dir/src/cpu.c.o \
+  CMakeFiles/bob16.dir/src/assembler.c.o \
   /usr/lib/gcc/x86_64-pc-linux-gnu/16/libgcc.a \
   /usr/lib/gcc/x86_64-pc-linux-gnu/16/libgcc_s_asneeded.so \
   /usr/lib/gcc/x86_64-pc-linux-gnu/16/libgcc_s_asneeded.so \
@@ -42,6 +44,10 @@ bob16: \
 /usr/lib/gcc/x86_64-pc-linux-gnu/16/crtbeginS.o:
 
 CMakeFiles/bob16.dir/src/main.c.o:
+
+CMakeFiles/bob16.dir/src/cpu.c.o:
+
+CMakeFiles/bob16.dir/src/assembler.c.o:
 
 /usr/lib/gcc/x86_64-pc-linux-gnu/16/libgcc.a:
