+++ build-asan/CMakeFiles/bob16.dir/link.txt
@@ -1,1 +1,1 @@
-/usr/bin/cc -Wl,--dependency-file=CMakeFiles/bob16.dir/link.d CMakeFiles/bob16.dir/src/main.c.o -o bob16
+/usr/bin/cc -Wl,--dependency-file=CMakeFiles/bob16.dir/link.d CMakeFiles/bob16.dir/src/main.c.o CMakeFiles/bob16.dir/src/cpu.c.o CMakeFiles/bob16.dir/src/assembler.c.o -o bob16
