+++ build-asan/CMakeFiles/bob16.dir/progress.make
@@ -1,3 +1,5 @@
 CMAKE_PROGRESS_1 = 1
 CMAKE_PROGRESS_2 = 2
+CMAKE_PROGRESS_3 = 3
+CMAKE_PROGRESS_4 = 4
