+++ include/assembler.h
@@ -1,0 +1,7 @@
+#ifndef ASSEMBLER_H
+#define ASSEMBLER_H
+
+int assemble(void);
+int parseReg(const char *s);
+
+#endif
